[ScyllaDll] Use handle structure to store all the necessary context infos
Scylla currently relies a lot on static variables for contextual
informations (target process, loaded modules, etc.) which is not great
in this case.

This commit introduce an opaque structure (SCY_HANDLE) which mirrors the
way Microsoft implements context struct through handles.

This is currently still wonky, but it's a work in progress.